# Conflicts:
#	app/src/main/res/layout/activity_hoabl_splash.xml

+++ networklayer/src/main/java/com/emproto/networklayer/response/portfolio/dashboard/Ongoing.kt
@@ -3,6 +3,6 @@ package com.emproto.networklayer.response.portfolio.dashboard
 data class Ongoing(
     val amountPaid: Double,
     val amountPending: Double,
-    val areaSqFt: Int,
+    val areaSqFt: Double,
     val count: Int
 )
